+++ app/Models/Identity/SocialAccount.php
@@ -43,6 +43,9 @@ class SocialAccount extends Model
         ];
     }
 
+    /**
+     * @return BelongsTo<User, $this>
+     */
     public function user(): BelongsTo
     {
         return $this->belongsTo(User::class);