+++ app/Models/Identity/User.php
@@ -711,6 +711,9 @@ class User extends Authenticatable implements HasMedia, MustVerifyEmail
         return $this->hasMany(UsernameChange::class);
     }
 
+    /**
+     * @return HasMany<SocialAccount, $this>
+     */
     public function socialAccounts(): HasMany
     {
         return $this->hasMany(SocialAccount::class);